[arcane,utils] Déplace la classe 'ExtentsV' dans son propre fichier.
